refactor: item list
