+++ src/components/comments/comments.jsx
@@ -7,6 +7,7 @@ import { useParams } from 'react-router';
 import { loadComments, updateComments } from '../../store/actions';
 import { useDispatch } from 'react-redux';
 import CommentList from './comment-list';
+import Loader from '../loader/loader';
 
 
 function Comments(){
@@ -14,62 +15,42 @@ function Comments(){
   const { id } = useParams();
   const dispatch = useDispatch();
   
-  const items = useSelector((state) => state.itemReducer);    
+  const items = useSelector((state) => state.itemReducer);
+  const { loading } = useSelector((state) => state.commentsReducer);
   const [commentsCount, setCommentsCount] = useState(0);
 
   useEffect(() => {
     if (!items[id]) return;
     setCommentsCount(items[id].kids.length);
     dispatch(loadComments(id, items[id].kids));
-  });
+  }, []);
 
   const handleReloadComments = () => {
     dispatch(updateComments(id));
     setCommentsCount(items[id].kids.length);
+    
   };
   
+  console.log(`loading`, loading);
   return(
     <Grid item xs={12}>
-      {
-        commentsCount ? (
-          <Grid container className={classes.boxWrapper}>
-
-            <Grid item xs={10}>
-              <Typography variant="h5"> Comments {commentsCount} </Typography>
-            </Grid>
-            <Grid item xs={2}>
-              <Button
-                variant="contained"
-                color="primary"
-                startIcon={<RefreshIcon />}
-                onClick={handleReloadComments}
-              > Refresh </Button>
-            </Grid>
-          
-            <Grid item xs={12}>
-              <CommentList />
-            </Grid>
-            
+      { loading ? <Loader /> : (
+        <Grid container className={classes.boxWrapper}>
+          <Grid item xs={10}>
+            <Typography variant="h5"> Comments {commentsCount} </Typography>
           </Grid>
-        ) : (
-          <Grid container className={classes.boxWrapper}>
-            <Grid item xs={12}>
-              <Grid container spacing={3} className={classes.header}>
-                <Grid item xs={10}>
-                  <Typography variant="h5"> Sorry! No comments posted so far. </Typography>
-                </Grid>
-                <Grid item xs={2}>
-                  <Button
-                    variant="contained"
-                    color="primary"
-                    startIcon={<RefreshIcon />}
-                    onClick={() => {}}
-                  > Refresh </Button>
-                </Grid>
-              </Grid>
-            </Grid>
-          </Grid>
-        )
+          <Grid item xs={2}>
+            <Button
+              variant="contained"
+              color="primary"
+              startIcon={<RefreshIcon />}
+              onClick={handleReloadComments}
+            > Refresh </Button>
+          </Grid>        
+          <Grid item xs={12}>
+            <CommentList />
+          </Grid>          
+        </Grid>)   
       }
     </Grid>
   );
