+++ src/components/item-list/item-list.jsx
@@ -1,25 +1,8 @@
 import { Grid } from '@material-ui/core';
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import Item from '../item/item';
 
-function ItemList(){
-  const [itemIds, setItemIds] = useState([]);
-
-  useEffect(() => {
-
-    const getTopStories = async () => {
-      const request = await fetch('https://hacker-news.firebaseio.com/v0/topstories.json?print=pretty');
-      return request.json();      
-    };
-    
-    getTopStories().then(data => {
-      const limitItemIds = data.slice(0, 10);
-      setItemIds(limitItemIds);
-    });
-
-  }, []);
-
-
+function ItemList({ itemIds }){
   return(
     <Grid container spacing={3}>
       {
