+++ src/store/actionTypes.js
@@ -3,3 +3,6 @@ export const ITEM_LOADER_ON = 'ITEM_LOADER_ON';
 export const ITEM_LOADER_OFF = 'ITEM_LOADER_OFF';
 
 export const LOAD_COMMENTS = 'LOAD_COMMENTS';
+
+export const COMMENTS_LOADER_ON = 'COMMENTS_LOADER_ON';
+export const COMMENTS_LOADER_OFF = 'COMMENTS_LOADER_OFF';
